add autograd support for matrix multiplication (#32)
This pr serves as a fix for #19, as well as providing testing coverage for the backwards pass and the forwards pass of sail.matmul

NOTE: this disables testing when using github actions. For some reason there is a segfault. Going to investigate with some clean docker images and replicate the build methods used in the action.